+++ frontend/src/css/components/InputSearch.css
@@ -1,0 +1,42 @@
+.searchInputContainer {
+  width: 100%;
+}
+
+.inputMessage {
+  font-size: 1rem;
+  display: flex;
+  justify-content: center;
+}
+
+.textInput.searchBar {
+  width: 100%;
+  border-radius: 6rem;
+  padding: 0.5rem;
+  border: solid 1px var(--basic-grey);
+  box-shadow: var(--shadow_initial);
+  background-image: url("../../assets/icons/search_black.svg");
+  background-repeat: no-repeat;
+  background-position: calc(100% - 95%) center;
+  padding-left: 3rem;
+  margin-bottom: 0;
+}
+
+@media screen and (min-width: 768px) {
+  .searchInputContainer{
+    display: flex;
+    justify-content: center;
+  }
+  .textInput.searchBar {
+    width: 60%;
+    background-position: calc(100% - 98%) center;
+    padding: 1rem 1rem 1rem 2.9rem;
+    &::placeholder {
+      font-size: 1.3rem;
+    }
+  }
+
+  .inputMessage {
+    font-size: 1.5rem;
+
+  }
+}
