+++ frontend/src/css/pages/Home.css
@@ -1,45 +1,29 @@
-.container_Home {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
 .imageContainer {
-  margin-top: 20px;
+  width: 100%;
+  height: 20rem;
 }
 
-.imageContainer img {
-  width: 100%;
-  height: 400px;
-  object-fit: contain;
-}
-
-.searchContainer {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.inputContainer {
+  padding: 0 2rem;
   position: relative;
-}
-.containerTextInput.searchBar {
-  max-width: 70%;
-  border-bottom: inherit;
+  width: 100%;
+  
 }
 
-.textInput.searchBar {
-  width: 70%;
-  border-radius: 6rem;
-  padding: 1rem;
-  padding-right: 30px;
-  border: solid 1px var(--basic-grey);
+.resultsSearch {
+  visibility: hidden;
 }
 
-.iconForm.searchBar {
+.resultsSearch.show {
+  visibility: visible;
+}
+
+.searchSuggestions {
   position: absolute;
-  right: 31%;
-  top: 50%;
-  transform: translateY(-50%);
+  z-index: 3;
+  width: 20rem;
+  background-color: #898989ce;
+  padding: 0.5rem;
 }
 
 .carouselContent {
@@ -72,10 +56,6 @@
   margin: 40px 0 0 30px;
 }
 
-.iconForm {
-  margin-right: 20px;
-}
-
 .carouselContainer {
   position: relative;
 }
@@ -95,63 +75,15 @@
   right: 10px;
 }
 
-/* recherche new */
-.searchContainer {
-  margin-top: 20px;
-  margin-bottom: 130px;
-  display: flex;
-  align-items: center;
-  position: relative;
-}
-
-.searchIcon {
-  position: absolute;
-  top: 50%;
-  right: 10px; /* Ajustez cette valeur en fonction de vos besoins */
-  transform: translateY(-50%);
-}
-/* gestien de suggestion recherche*/
-.searchContainer {
-  position: relative;
-}
-
-.searchSuggestionsContainer {
-  position: absolute;
-  top: calc(100% + 5px);
-  /* Ajustez la valeur pour définir l'espace entre l'input et les suggestions */
-  width: 100%;
-  background-color: #fff; /* Couleur de fond des suggestions */
-  border-radius: 5px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Ombre des suggestions */
-}
-
-.searchSuggestions {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.searchSuggestions li {
-  padding: 10px;
-  cursor: pointer;
-}
-
-.searchSuggestions li:hover {
-  background-color: #f5f5f5; /* Couleur de fond au survol d'une suggestion */
-}
-.searchSuggestions li.selected {
-  background-color: #f5f5f5;
-}
-
-@media screen and (max-width: 576px) {
-  .searchInput {
-    width: 100%;
+@media screen and (min-width: 768px) {
+  .imageContainer {
+    height: 35rem;
   }
-
-  .imageContainer img {
-    height: 300px;
+  .inputContainer {
+   display: flex;
+   flex-direction: column;
+   margin: 0 auto;
   }
-
   .carouselItem {
     flex: 0 0 100%;
   }
